check-before-push.sh: support for specifying the FORMULA name in the second argument

+++ check-before-push.sh
@@ -7,6 +7,7 @@ set -x
 
 # local TAP (brew tap)
 TAP="$1"
+FORMULA="${2:-}"
 
 export HOMEBREW_NO_AUTO_UPDATE=1
 
@@ -16,6 +17,10 @@ for filepath in Formula/*.rb
 do
   fname="${filepath##*/}"
   formula="${fname%.rb}"
+  if [[ -n "${FORMULA}" ]] && [[ "${FORMULA}" != "${formula}" ]]
+  then
+    continue
+  fi
   brew uninstall "${formula}" || echo "(IGNORED)"
 done
 
@@ -23,6 +28,10 @@ for filepath in Formula/*.rb
 do
   fname="${filepath##*/}"
   formula="${fname%.rb}"
+  if [[ -n "${FORMULA}" ]] && [[ "${FORMULA}" != "${formula}" ]]
+  then
+    continue
+  fi
   brew install --build-from-source "${TAP}/${formula}"
   brew audit --formula "${TAP}/${formula}"
   brew test "${TAP}/${formula}"
